split into two tests
