+++ test/e2e/app-dir/middleware-sitemap/no-matcher/app/layout.tsx
@@ -1,0 +1,7 @@
+export default function Root({ children }) {
+  return (
+    <html>
+      <body>{children}</body>
+    </html>
+  )
+}
