+++ test/e2e/app-dir/middleware-sitemap/no-matcher/app/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <p>hello world</p>
+}
