+++ test/e2e/app-dir/middleware-sitemap/no-matcher/app/redirected/page.tsx
@@ -1,0 +1,3 @@
+export default function Redirected() {
+  return <p>redirected</p>
+}
