+++ test/e2e/app-dir/middleware-sitemap/no-matcher/app/sitemap.ts
@@ -1,0 +1,12 @@
+import type { MetadataRoute } from 'next'
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  return [
+    {
+      url: 'https://vercel.com',
+      lastModified: new Date(),
+      changeFrequency: 'yearly',
+      priority: 1,
+    },
+  ]
+}
