+++ test/e2e/app-dir/middleware-sitemap/no-matcher/index.test.ts
@@ -1,0 +1,15 @@
+import { nextTestSetup } from 'e2e-utils'
+
+describe('middleware-sitemap', () => {
+  const { next } = nextTestSetup({
+    files: __dirname,
+  })
+
+  it('should be affected by middleware for sitemap.xml if there is no matcher', async () => {
+    let html = await next.render('/')
+    expect(html).toContain('redirected')
+
+    html = await next.render('/sitemap.xml')
+    expect(html).toContain('redirected')
+  })
+})
